requirements.txt
pandas
fastapi
requests
gunicorn
spacy
uvicorn
numpy
python-decouple
scikit-learn
pydantic-settings
motor
pydantic[email]
python-jose
firebase-admin
passlib
pyjwt
bcrypt
python-multipart
httpx